removed fileExists in favor of a try/catch; cleaned up whitespace
checking to see if file exists before loading it, is supposedly an
anti-pattern. While it worked and didn’t break my tests, for some
reason I’m not sure about, my test failed in travis-ci. I just removed
the fileExists check and have just wrapped that specific code in a
try/catch and letting the code to continue